Completed the draw functions
Draw functions are basiclly completed. The text can be moved,
to move it, just click upper conner of it and dragged the mouse.

The drawing function has a bug in a very special condition, but I don't know why is that and when will it happen. So just try as much as possiable operations to find it and tell me.

The new session is not perfect,  once the first window closed, all other window closed again.

The new tab function is not complete, I really think this version can not use tab to a new session, otherwise, the program should rewrite much.